Refactor Tooltip to conform to MVC arch
